# server/docs/swagger.yaml
definitions:
...
  huasu.CarouselManage:
    properties:
      ID:
        description: 主键ID
        type: integer
      alt:
        description: 图片的 alt 属性
        type: string
      business:
        description: 关联业务模块
        type: string
      created_by:
        description: 创建人 ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      description:
        description: 轮播图描述（管理后台备注，如 “618 活动主视觉”）
        type: string
      image:
        description: 轮播图片的存储路径或 CDN 链接
        type: string
      link:
        description: 点击轮播图后的跳转链接
        type: string
      sort:
        description: 排序权重
        type: integer
      status:
        description: 状态（如：1 - 启用、0 - 禁用、2 - 草稿，控制是否在前端显示）
        type: integer
      target:
        description: 链接打开方式
        type: string
      title:
        description: 轮播图标题
        type: string
      updatedAt:
        description: 更新时间
        type: string
    required:
    - business
    - image
    - link
    - sort
    - status
    - title
    type: object
  huasu.Notice:
    properties:
      ID:
        description: 主键ID
        type: integer
      content:
        description: 公告正文（必填）
        type: string
      cover:
        description: 封面图 URL（可选）
        type: string
      createdAt:
        description: 创建时间
        type: string
      end_time:
        description: 失效时间（可选，自动下架）
        type: string
      source:
        description: 公告来源（可选）
        type: string
      start_time:
        description: 生效时间（可选，定时发布）
        type: string
      status:
        description: 发布状态（必填），控制公告是否对外展示，常见值： - 0：草稿（仅创建者可见，未提交审核） - 1：待审核（提交给管理员，未发布）
          - 2：已发布（全员 / 指定用户可见） - 3：已下架（历史公告，不再展示）
        type: integer
      summary:
        description: 公告摘要（可选）
        type: string
      title:
        description: 公告标题（必填）
        type: string
      top:
        description: 是否置顶（可选），1 = 置顶，0 = 不置顶，置顶公告在列表页优先排序（如紧急系统通知）。
        type: integer
      type:
        description: 公告类型（必填，用于分类筛选），如： - 按业务分：系统公告、活动公告、通知公告、预警公告 - 按部门分：全员公告、技术部公告、财务部公告
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    required:
    - content
    - status
    - title
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.5
paths:
  /Carousel/createCarouselManage:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建轮播图管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/huasu.CarouselManage'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建轮播图管理
      tags:
      - CarouselManage
  /Carousel/deleteCarouselManage:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除轮播图管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/huasu.CarouselManage'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除轮播图管理
      tags:
      - CarouselManage
  /Carousel/deleteCarouselManageByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除轮播图管理
      tags:
      - CarouselManage
  /Carousel/findCarouselManage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询轮播图管理
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/huasu.CarouselManage'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询轮播图管理
      tags:
      - CarouselManage
  /Carousel/getCarouselManageList:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: createdAtRange
        type: array
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: order
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取轮播图管理列表
      tags:
      - CarouselManage
...
  /Carousel/updateCarouselManage:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新轮播图管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/huasu.CarouselManage'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新轮播图管理
      tags:
      - CarouselManage
...
  /notice/createNotice:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建公告管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/huasu.Notice'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建公告管理
      tags:
      - Notice
  /notice/deleteNotice:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除公告管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/huasu.Notice'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除公告管理
      tags:
      - Notice
  /notice/deleteNoticeByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除公告管理
      tags:
      - Notice
  /notice/findNotice:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询公告管理
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/huasu.Notice'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询公告管理
      tags:
      - Notice
  /notice/getNoticeList:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: createdAtRange
        type: array
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取公告管理列表
      tags:
      - Notice
...
  /notice/updateNotice:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新公告管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/huasu.Notice'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新公告管理
      tags:
      - Notice
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...